@@ -1,5 +1,8 @@
 <?php
 
+declare(strict_types=1);
+
+namespace Tests\Support;
 
 /**
  * Inherited Methods
@@ -12,15 +15,15 @@
  * @method void am($role)
  * @method void lookForwardTo($achieveValue)
  * @method void comment($description)
- * @method \Codeception\Lib\Friend haveFriend($name, $actorClass = NULL)
+ * @method void pause($vars = [])
  *
  * @SuppressWarnings(PHPMD)
 */
-class UnitGuy extends \Codeception\Actor
+class FunctionalTester extends \Codeception\Actor
 {
-    use _generated\UnitGuyActions;
+    use _generated\FunctionalTesterActions;
 
-   /**
-    * Define custom actions here
-    */
+    /**
+     * Define custom actions here
+     */
 }
